+++ HISTORY.md
@@ -1,6 +1,13 @@
 History
 =======
 
+0.0.2 (2013-8-09)
+-----------------
+
+- improved security - ask for password in browser
+- added interactive mode that keeps browser open
+- fixed timeout when "allow all javascript" mode is turned on
+
 0.0.1 (2013-7-25)
 -----------------